Sistema com relatorios funcionais com tag selector

+++ ControleFutebolWeb/Models/Notadetalhe.cs
@@ -9,6 +9,6 @@
         public string AcaoId { get; set; }    // ex: "drible_certo"
         public string AcaoLabel { get; set; } // ex: "Drible bem-sucedido"
         public int Quantidade { get; set; }   // quantas vezes clicou
-        public int Peso { get; set; }         // +1 ou -1
+        public double Peso { get; set; }      // ex: +0.1, -0.5, +2
     }
 }
